server/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Redis==0.4.0
Flask-SocketIO==5.3.6
eventlet==0.36.1
paho-mqtt==2.1.0
PyMySQL==1.1.1
